requirements.txt
asgiref==3.3.4
autopep8==1.5.4
certifi==2020.12.5
Django==3.2
pycodestyle==2.6.0
pytz==2021.1
sqlparse==0.4.1
wincertstore==0.2
pillow==8.2.0
django-ckeditor=6.0.0
